Added enough infra to ZeroDynVector to allow creating, appending and writing dynvectors. Now the tools write out gkyl format dynvectors and beautiful plots can be made of speratrix shape, for example, from eqdsk files etc.

+++ zero/dynvec.c
@@ -56,7 +56,7 @@ gkyl_dynvec_new(enum gkyl_elem_type type, size_t ncomp)
 }
 
 int gkyl_dynvec_elem_type(gkyl_dynvec vec) { return vec->type; }
-size_t gkyl_dynvec_ncomp(gkyl_dynvec vec) { return vec->ncomp; }
+int gkyl_dynvec_ncomp(gkyl_dynvec vec) { return vec->ncomp; }
 
 void
 gkyl_dynvec_reserve_more(gkyl_dynvec dv, size_t rsize)
